[SPARK-46500][PS][TESTS] Reorganize `FrameParityPivotTests`
### What changes were proposed in this pull request?
Reorganize `FrameParityPivotTests`: break `test_pivot_table` into mutiple tests

### Why are the changes needed?
this test is slow
```
Starting test(python3.9): pyspark.pandas.tests.connect.computation.test_parity_pivot (temp output: /__w/spark/spark/python/target/5f37e442-9037-47cc-8c6b-e9a273299d0d/python3.9__pyspark.pandas.tests.connect.computation.test_parity_pivot__ozvdx_ay.log)
Finished test(python3.9): pyspark.pandas.tests.connect.computation.test_parity_pivot (524s)
```

### Does this PR introduce _any_ user-facing change?
no, test only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44478 from zhengruifeng/ps_test_pivot_multi.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
